fix state bug with instagram post
